[backend] test coverage sco domain (#10905)
